Moved ExitCode and ExitCodesNamespace to common
