+++ aiida/backends/tests/engine/test_work_chain.py
@@ -24,7 +24,8 @@ from aiida.backends.testbase import AiidaTestCase
 from aiida.common import exceptions
 from aiida.common.links import LinkType
 from aiida.common.utils import Capturing
-from aiida.engine import ExitCode, Process, ToContext, WorkChain, if_, while_, return_, launch, calcfunction
+from aiida.common.exit_codes import ExitCode
+from aiida.engine import Process, ToContext, WorkChain, if_, while_, return_, launch, calcfunction
 from aiida.engine.persistence import ObjectLoader
 from aiida.manage.manager import get_manager
 from aiida.orm import load_node, Bool, Float, Int, Str
