+++ aiida/engine/processes/__init__.py
@@ -13,15 +13,15 @@ from __future__ import division
 from __future__ import print_function
 from __future__ import absolute_import
 
+from aiida.common.exit_codes import *
+
 from .builder import *
 from .calcjobs import *
-from .exit_code import *
 from .functions import *
 from .ports import *
 from .process import *
 from .workchains import *
 
 __all__ = (
-    builder.__all__ + calcjobs.__all__ + exit_code.__all__ + functions.__all__ + ports.__all__ + process.__all__ +
-    workchains.__all__
+    builder.__all__ + calcjobs.__all__ + functions.__all__ + ports.__all__ + process.__all__ + workchains.__all__
 )
